--- web/frontend/src/pages/session/Login.css
@@ -1,20 +1,0 @@
-.login-container {
-  min-width: 58ch;
-  display: flex;
-  text-align: center;
-  flex-direction: column;
-  padding-top: 2%;
-  padding-left: 20%;
-  padding-right: 20%;
-  justify-content: center;
-  background-color: whitesmoke;
-  padding-bottom: 200px;
-}
-.login-txt {
-  display: flex;
-  padding-top: 5%;
-  padding-bottom: 10%;
-}
-.login-btn {
-  display: flex;
-}
